+++ examples/CMakeLists.txt
@@ -1,5 +1,3 @@
 add_subdirectory(_shared)
 add_subdirectory(har_classifier)
-add_subdirectory(har_classifier_v2)
 add_subdirectory(ecg_anomaly_ae)
-add_subdirectory(ecg_anomaly_ae_v2)
