+++ examples/ecg_anomaly_ae/train_c.c
@@ -8,24 +8,24 @@
 #include <sys/stat.h>
 #include <time.h>
 
-#include "AvgPool1d.h"
 #include "CalculateGradsSequential.h"
 #include "Common.h"
 #include "Conv1dApi.h"
-#include "Conv1dTransposed.h" /* no userApi yet — manual build below */
+#include "Conv1dTransposedApi.h"
 #include "DataLoader.h"
 #include "DataLoaderApi.h"
-#include "Distributions.h"
 #include "InferenceApi.h"
-#include "Kernel.h"
 #include "Layer.h"
+#include "LayerCommon.h"
+#include "LayerQuant.h"
 #include "LossFunction.h"
-#include "MaxPool1d.h"
 #include "NPYLoaderApi.h"
+#include "Pool1dApi.h"
 #include "Quantization.h"
 #include "QuantizationApi.h"
 #include "ReluApi.h"
 #include "SgdApi.h"
+#include "StateDictApi.h"
 #include "StorageApi.h"
 #include "Tensor.h"
 #include "TensorApi.h"
@@ -43,14 +43,16 @@
 #define IN_CHANNELS 1
 #define LEN_INPUT 140
 
-/* Encoder channel widths */
 #define E1_OUT 8
 #define E1_K 7
 #define E1_S 2
+/* enc1 is a stride-2 conv; PyTorch trained it with symmetric padding=3. C SAME
+ * would pick the minimal/asymmetric pad {2,3} and diverge, so use EXPLICIT
+ * padding=(K-1)/2=3 to match PyTorch bit-for-bit (issue #177). */
+#define E1_PAD (E1_K / 2)
 #define E2_OUT 16
 #define E2_K 5
 
-/* Decoder channel widths and kernel/strides (K=2,S=2 substitution for K=4-pad=1 spec) */
 #define D1_OUT 8
 #define D1_K 5
 #define D1_S 5
@@ -61,134 +63,12 @@
 #define D3_K 2
 #define D3_S 2
 
-/* Encoder: 2× (Conv1d + ReLU + Pool) = 6 layers
- * Decoder: 3× ConvT1d + 2× ReLU = 5 layers
- * Total = 11 */
 #define MODEL_SIZE 11
-
-/* Forward declaration; defined in Task 6. */
-static void buildModel(layer_t **model);
-
-/* ------------------------------------------------------------------------- */
-/* Model parameters (file-static — must outlive buildModel).                 */
-/* ------------------------------------------------------------------------- */
-
-/* Conv1d weights: [Cout, Cin, K]. Bias: [Cout] rank-1 (matches Conv1d.c). */
-static float e1_w_data[E1_OUT * IN_CHANNELS * E1_K];
-static size_t e1_w_dims[3] = {E1_OUT, IN_CHANNELS, E1_K};
-static float e1_b_data[E1_OUT];
-static size_t e1_b_dims[1] = {E1_OUT};
-
-static float e2_w_data[E2_OUT * E1_OUT * E2_K];
-static size_t e2_w_dims[3] = {E2_OUT, E1_OUT, E2_K};
-static float e2_b_data[E2_OUT];
-static size_t e2_b_dims[1] = {E2_OUT};
-
-/* Conv1dTransposed weights: [Cin, Cout/groups, K]  (note the SWAP from Conv1d).
- * Per src/layer/include/Conv1dTransposed.h:14. Bias: [Cout] rank-1. */
-static float d1_w_data[E2_OUT * D1_OUT * D1_K];
-static size_t d1_w_dims[3] = {E2_OUT, D1_OUT, D1_K};
-static float d1_b_data[D1_OUT];
-static size_t d1_b_dims[1] = {D1_OUT};
-
-static float d2_w_data[D1_OUT * D2_OUT * D2_K];
-static size_t d2_w_dims[3] = {D1_OUT, D2_OUT, D2_K};
-static float d2_b_data[D2_OUT];
-static size_t d2_b_dims[1] = {D2_OUT};
-
-static float d3_w_data[D2_OUT * D3_OUT * D3_K];
-static size_t d3_w_dims[3] = {D2_OUT, D3_OUT, D3_K};
-static float d3_b_data[D3_OUT];
-static size_t d3_b_dims[1] = {D3_OUT};
-
-static parameter_t *buildParam(distributionType_t dist, float *data, size_t *dims, size_t ndim,
-                               size_t fanIn, size_t fanOut) {
-    quantization_t *q = quantizationInitFloat();
-    tensor_t *p = tensorInitWithDistribution(dist, data, dims, ndim, q, NULL, fanIn, fanOut);
-    tensor_t *g = gradInitFloat(p, NULL);
-    return parameterInit(p, g);
-}
-
-static layer_t *buildMaxPool1dLayer(size_t kSize, size_t stride, size_t outC, size_t outLen) {
-    quantization_t *q = quantizationInitFloat();
-
-    kernel_t *kernel = reserveMemory(sizeof(kernel_t));
-    initKernel(kernel, kSize, VALID, /*dilation*/ 1, stride);
-
-    /* Argmax buffer is sized for B=1 (training_batch iterates microbatch-by-
-     * microbatch), shape [1, outC, outLen]. */
-    size_t numArgmax = 1 * outC * outLen;
-    int32_t *argmaxBuf = reserveMemory(numArgmax * sizeof(int32_t));
-    size_t *argmaxDims = reserveMemory(3 * sizeof(size_t));
-    argmaxDims[0] = 1;
-    argmaxDims[1] = outC;
-    argmaxDims[2] = outLen;
-    tensor_t *argmax = tensorInitInt32(argmaxBuf, argmaxDims, 3, NULL);
-
-    maxPool1dConfig_t *cfg = reserveMemory(sizeof(maxPool1dConfig_t));
-    initMaxPool1dConfig(cfg, kernel, argmax, q, q);
-
-    layer_t *layer = reserveMemory(sizeof(layer_t));
-    layerConfig_t *lc = reserveMemory(sizeof(layerConfig_t));
-    layer->type = MAXPOOL1D;
-    lc->maxPool1d = cfg;
-    layer->config = lc;
-    return layer;
-}
-
-static layer_t *buildAvgPool1dLayer(size_t kSize, size_t stride) {
-    quantization_t *q = quantizationInitFloat();
-
-    kernel_t *kernel = reserveMemory(sizeof(kernel_t));
-    initKernel(kernel, kSize, VALID, /*dilation*/ 1, stride);
-
-    avgPool1dConfig_t *cfg = reserveMemory(sizeof(avgPool1dConfig_t));
-    initAvgPool1dConfig(cfg, kernel, q, q);
-
-    layer_t *layer = reserveMemory(sizeof(layer_t));
-    layerConfig_t *lc = reserveMemory(sizeof(layerConfig_t));
-    layer->type = AVGPOOL1D;
-    lc->avgPool1d = cfg;
-    layer->config = lc;
-    return layer;
-}
-
-/* Conv1dTransposed has no userApi yet (Phase 1 contract: paddingType_t = VALID
- * mandatory; SAME is rejected with PRINT_ERROR + exit). We mirror the manual
- * idiom from test/unit/layer/UnitTestConv1dTransposed.c, but use reserveMemory
- * so the cfg/layer survive across multiple buildModel calls (which doesn't
- * happen here, but is consistent with the rest of the file). */
-static layer_t *buildConv1dTransposedLayer(parameter_t *w, parameter_t *b, size_t kSize,
-                                           size_t stride, size_t outputPadding, size_t groups) {
-    quantization_t *q = quantizationInitFloat();
-
-    kernel_t *kernel = reserveMemory(sizeof(kernel_t));
-    initKernel(kernel, kSize, VALID, /*dilation*/ 1, stride);
-
-    conv1dTransposedConfig_t *cfg = reserveMemory(sizeof(conv1dTransposedConfig_t));
-    initConv1dTransposedConfigWithWeightsAndBias(cfg, kernel, w, b, groups, outputPadding, q, q, q,
-                                                 q);
-
-    layer_t *layer = reserveMemory(sizeof(layer_t));
-    layerConfig_t *lc = reserveMemory(sizeof(layerConfig_t));
-    layer->type = CONV1D_TRANSPOSED;
-    lc->conv1dTransposed = cfg;
-    layer->config = lc;
-    return layer;
-}
-
-/* ------------------------------------------------------------------------- */
-/* Datasets and dataloader thunks.                                           */
-/* ------------------------------------------------------------------------- */
 
 static dataset_t g_trainDataset;
 static dataset_t g_valDataset;
 static dataset_t g_testDataset;
 
-/* npyLoad strips the leading N dim, leaving each item with shape [1, 140]
- * rank-2. The C model expects rank-3 inputs [B=1, 1, 140] for Conv1d. The MSE
- * loss expects the label to have the same shape as the model output. Both
- * items AND labels are reshaped to [1, 1, 140]. */
 static void reshapeItemsAddBatchDim(tensorArray_t *items) {
     for (size_t i = 0; i < items->size; ++i) {
         tensor_t *t = items->array[i];
@@ -213,9 +93,6 @@ static void reshapeItemsAddBatchDim(tensorArray_t *items) {
     }
 }
 
-/* AE: label IS the input. We re-load the same .npy file as the label tensor.
- * Two npyLoad calls produce two independent copies (no aliasing); RAM cost is
- * trivial (≤ 200 KB doubled for ECG5000). */
 static void initDataSets(void) {
     tensorArray_t *trainItems = npyLoad("examples/ecg_anomaly_ae/data/train_x.npy");
     tensorArray_t *trainLabels = npyLoad("examples/ecg_anomaly_ae/data/train_x.npy");
@@ -248,7 +125,6 @@ static sample_t *getValSample(size_t id) {
 static sample_t *getTestSample(size_t id) {
     return npyGetSample(&g_testDataset, id);
 }
-
 static size_t getTrainSize(void) {
     return g_trainDataset.items->size;
 }
@@ -259,78 +135,94 @@ static size_t getTestSize(void) {
     return g_testDataset.items->size;
 }
 
-static void buildModel(layer_t **model) {
-    quantization_t *q = quantizationInitFloat();
+static void buildModel(layer_t **model, layerQuant_t *lq) {
+    /* Encoder */
+    model[0] = conv1dLayerInit(&(conv1dInit_t){.inChannels = IN_CHANNELS,
+                                               .outChannels = E1_OUT,
+                                               .kernelSize = E1_K,
+                                               .stride = E1_S,
+                                               .padding = EXPLICIT,
+                                               .paddingAmount = E1_PAD},
+                               lq);
+    model[1] = reluLayerInit(lq);
+    model[2] = maxPool1dLayerInit(
+        &(maxPool1dInit_t){
+            .kernelSize = 2, .stride = 2, .inputChannels = E1_OUT, .inputLength = LEN_INPUT / E1_S},
+        lq);
 
-    /* ---- Encoder ---- */
+    model[3] = conv1dLayerInit(
+        &(conv1dInit_t){
+            .inChannels = E1_OUT, .outChannels = E2_OUT, .kernelSize = E2_K, .padding = SAME},
+        lq);
+    model[4] = reluLayerInit(lq);
+    model[5] = avgPool1dLayerInit(&(avgPool1dInit_t){.kernelSize = 5, .stride = 5}, lq);
 
-    /* Block E1: Conv1d(1→8, K=7, S=2, padding=SAME), ReLU.
-     * SAME with stride=2 on len 140 → len 70. */
-    kernel_t *e1k = reserveMemory(sizeof(kernel_t));
-    initKernel(e1k, E1_K, SAME, /*dilation*/ 1, /*stride*/ E1_S);
-    parameter_t *e1_w =
-        buildParam(XAVIER_UNIFORM, e1_w_data, e1_w_dims, 3, IN_CHANNELS * E1_K, E1_OUT * E1_K);
-    parameter_t *e1_b = buildParam(ZEROS, e1_b_data, e1_b_dims, 1, 1, E1_OUT);
-    model[0] = conv1dLayerInitLegacy(e1_w, e1_b, e1k, q, q, q, q);
-    model[1] = reluLayerInitLegacy(quantizationInitFloat(), quantizationInitFloat());
+    /* Decoder */
+    model[6] = conv1dTransposedLayerInit(
+        &(conv1dTransposedInit_t){
+            .inChannels = E2_OUT, .outChannels = D1_OUT, .kernelSize = D1_K, .stride = D1_S},
+        lq);
+    model[7] = reluLayerInit(lq);
 
-    /* Block P1: MaxPool1d(K=2, S=2). 70 → 35. */
-    model[2] = buildMaxPool1dLayer(/*K*/ 2, /*S*/ 2, /*outC*/ E1_OUT, /*outLen*/ 35);
+    model[8] = conv1dTransposedLayerInit(
+        &(conv1dTransposedInit_t){
+            .inChannels = D1_OUT, .outChannels = D2_OUT, .kernelSize = D2_K, .stride = D2_S},
+        lq);
+    model[9] = reluLayerInit(lq);
 
-    /* Block E2: Conv1d(8→16, K=5, padding=SAME), ReLU. */
-    kernel_t *e2k = reserveMemory(sizeof(kernel_t));
-    initKernel(e2k, E2_K, SAME, 1, 1);
-    parameter_t *e2_w =
-        buildParam(XAVIER_UNIFORM, e2_w_data, e2_w_dims, 3, E1_OUT * E2_K, E2_OUT * E2_K);
-    parameter_t *e2_b = buildParam(ZEROS, e2_b_data, e2_b_dims, 1, 1, E2_OUT);
-    model[3] =
-        conv1dLayerInitLegacy(e2_w, e2_b, e2k, quantizationInitFloat(), quantizationInitFloat(),
-                              quantizationInitFloat(), quantizationInitFloat());
-    model[4] = reluLayerInitLegacy(quantizationInitFloat(), quantizationInitFloat());
-
-    /* Block P2: AvgPool1d(K=5, S=5). 35 → 7 (bottleneck). */
-    model[5] = buildAvgPool1dLayer(/*K*/ 5, /*S*/ 5);
-
-    /* ---- Decoder ---- */
-
-    /* Block D1: Conv1dTransposed(16→8, K=5, S=5, op=0). 7 → 35. ReLU. */
-    parameter_t *d1_w =
-        buildParam(XAVIER_UNIFORM, d1_w_data, d1_w_dims, 3, E2_OUT * D1_K, D1_OUT * D1_K);
-    parameter_t *d1_b = buildParam(ZEROS, d1_b_data, d1_b_dims, 1, 1, D1_OUT);
-    model[6] = buildConv1dTransposedLayer(d1_w, d1_b, /*K*/ D1_K, /*S*/ D1_S,
-                                          /*outputPadding*/ 0, /*groups*/ 1);
-    model[7] = reluLayerInitLegacy(quantizationInitFloat(), quantizationInitFloat());
-
-    /* Block D2: Conv1dTransposed(8→4, K=2, S=2, op=0). 35 → 70. ReLU. */
-    parameter_t *d2_w =
-        buildParam(XAVIER_UNIFORM, d2_w_data, d2_w_dims, 3, D1_OUT * D2_K, D2_OUT * D2_K);
-    parameter_t *d2_b = buildParam(ZEROS, d2_b_data, d2_b_dims, 1, 1, D2_OUT);
-    model[8] = buildConv1dTransposedLayer(d2_w, d2_b, /*K*/ D2_K, /*S*/ D2_S,
-                                          /*outputPadding*/ 0, /*groups*/ 1);
-    model[9] = reluLayerInitLegacy(quantizationInitFloat(), quantizationInitFloat());
-
-    /* Block D3: Conv1dTransposed(4→1, K=2, S=2, op=0). 70 → 140. NO ReLU on final. */
-    parameter_t *d3_w =
-        buildParam(XAVIER_UNIFORM, d3_w_data, d3_w_dims, 3, D2_OUT * D3_K, D3_OUT * D3_K);
-    parameter_t *d3_b = buildParam(ZEROS, d3_b_data, d3_b_dims, 1, 1, D3_OUT);
-    model[10] = buildConv1dTransposedLayer(d3_w, d3_b, /*K*/ D3_K, /*S*/ D3_S,
-                                           /*outputPadding*/ 0, /*groups*/ 1);
+    model[10] = conv1dTransposedLayerInit(
+        &(conv1dTransposedInit_t){
+            .inChannels = D2_OUT, .outChannels = D3_OUT, .kernelSize = D3_K, .stride = D3_S},
+        lq);
 }
 
-/* ------------------------------------------------------------------------- */
-/* Per-epoch JSON log writer + epoch callback.                               */
-/* ------------------------------------------------------------------------- */
+static int loadStateDictFromDir(layer_t **model, const char *weightsDir) {
+    /* Param layer order in model[]: e1 (0), e2 (3), d1 (6), d2 (8), d3 (10). 5 entries. */
+    char wPath[256], bPath[256];
+    const char *names[5] = {"e1", "e2", "d1", "d2", "d3"};
+    tensor_t *w[5] = {0};
+    tensor_t *b[5] = {0};
+
+    for (int i = 0; i < 5; i++) {
+        snprintf(wPath, sizeof(wPath), "%s/%s.weight.npy", weightsDir, names[i]);
+        snprintf(bPath, sizeof(bPath), "%s/%s.bias.npy", weightsDir, names[i]);
+        /* npyLoadFlat (not npyLoad): a weight file is ONE tensor of shape
+         * [out, in, k] (Conv1d) or [in, out, k] (ConvTranspose1d). npyLoad()
+         * slices dim0 into row tensors, so array[0] is only the first channel;
+         * the subsequent layerLoadWeights memcpy then runs past that short
+         * buffer into heap garbage — the issue #177 collapse. */
+        w[i] = npyLoadFlat(wPath);
+        b[i] = npyLoadFlat(bPath);
+        if (w[i] == NULL || b[i] == NULL) {
+            fprintf(stderr, "loadStateDictFromDir: missing %s or %s\n", wPath, bPath);
+            return 1;
+        }
+    }
+
+    modelLoadStateDict(
+        model, MODEL_SIZE,
+        (stateDictEntry_t[]){
+            {.name = names[0], .weightData = (float *)w[0]->data, .biasData = (float *)b[0]->data},
+            {.name = names[1], .weightData = (float *)w[1]->data, .biasData = (float *)b[1]->data},
+            {.name = names[2], .weightData = (float *)w[2]->data, .biasData = (float *)b[2]->data},
+            {.name = names[3], .weightData = (float *)w[3]->data, .biasData = (float *)b[3]->data},
+            {.name = names[4], .weightData = (float *)w[4]->data, .biasData = (float *)b[4]->data},
+        },
+        5);
+
+    /* modelLoadStateDict copied the data into the layers; release the loaders. */
+    for (int i = 0; i < 5; i++) {
+        freeTensor(w[i]);
+        freeTensor(b[i]);
+    }
+    return 0;
+}
 
 static FILE *g_log_file = NULL;
 static int g_first_epoch = 1;
 static struct timespec g_epoch_t0;
 
 static void epochCallback(size_t epoch, float trainLoss, epochStats_t evalStats) {
-    /* trainingRun's eval pass derives numClasses from label_num_elements (140
-     * for our AE), so evalStats.accuracy / .precision / .recall / .f1 contain
-     * argmax-based 140-class noise. We drop them; only evalStats.loss is
-     * meaningful (it's the MSE-mean-per-element, matching PyTorch). val_acc
-     * is null in the JSON to match the PyTorch side. */
     struct timespec t1;
     clock_gettime(CLOCK_MONOTONIC, &t1);
     double wall_s =
@@ -353,10 +245,6 @@ static void epochCallback(size_t epoch, float trainLoss, epochStats_t evalStats)
     clock_gettime(CLOCK_MONOTONIC, &g_epoch_t0);
 }
 
-/* Run forward inference on every sample of the given dataset, allocate a
- * single contiguous [N, 1, 140] float buffer, fill it, and write it to
- * `outPath` via npyWriteFloat32. The buffer is malloc-owned and freed by
- * this function. */
 static int writeAllReconstructions(layer_t **model, size_t modelSize,
                                    sample_t *(*getSample)(size_t), size_t n, const char *outPath) {
     size_t totalElems = n * IN_CHANNELS * LEN_INPUT;
@@ -402,73 +290,79 @@ int main(void) {
 
     initDataSets();
 
-    dataLoader_t *trainLoader = dataLoaderInit(getTrainSample, getTrainSize, BATCH, NULL, NULL,
-                                               /*shuffle*/ true, /*shuffleSeed*/ SHUFFLE_SEED,
-                                               /*dropLast*/ true);
-    dataLoader_t *valLoader = dataLoaderInit(getValSample, getValSize, 1, NULL, NULL,
-                                             /*shuffle*/ false, /*shuffleSeed*/ 0,
-                                             /*dropLast*/ true);
     dataLoader_t *testLoader = dataLoaderInit(getTestSample, getTestSize, 1, NULL, NULL,
                                               /*shuffle*/ false, /*shuffleSeed*/ 0,
                                               /*dropLast*/ true);
 
+    layerQuant_t lq;
+    layerQuantInitUniform(&lq, quantizationInitFloat());
+
     layer_t *model[MODEL_SIZE];
-    buildModel(model);
+    buildModel(model, &lq);
 
-    optimizer_t *sgd =
-        sgdMCreateOptim(LR, MOMENTUM, /*weightDecay*/ 0.0f, model, MODEL_SIZE, FLOAT32);
+    const char *bitParity = getenv("BIT_PARITY");
+    if (bitParity != NULL && bitParity[0] != '\0') {
+        const char *wDir = "examples/ecg_anomaly_ae/weights";
+        if (loadStateDictFromDir(model, wDir) != 0) {
+            fprintf(stderr, "BIT_PARITY: state_dict load failed\n");
+            return 1;
+        }
+        fprintf(stdout, "BIT_PARITY: loaded state_dict from %s\n", wDir);
+    } else {
+        dataLoader_t *trainLoader = dataLoaderInit(getTrainSample, getTrainSize, BATCH, NULL, NULL,
+                                                   /*shuffle*/ true, /*shuffleSeed*/ SHUFFLE_SEED,
+                                                   /*dropLast*/ true);
+        dataLoader_t *valLoader = dataLoaderInit(getValSample, getValSize, 1, NULL, NULL,
+                                                 /*shuffle*/ false, /*shuffleSeed*/ 0,
+                                                 /*dropLast*/ true);
 
-    g_log_file = fopen("examples/ecg_anomaly_ae/logs/c.json", "w");
-    if (!g_log_file) {
-        fprintf(stderr, "ERROR: cannot open log file for writing\n");
-        return 1;
+        optimizer_t *sgd =
+            sgdMCreateOptim(LR, MOMENTUM, /*weightDecay*/ 0.0f, model, MODEL_SIZE, FLOAT32);
+
+        g_log_file = fopen("examples/ecg_anomaly_ae/logs/c.json", "w");
+        if (!g_log_file) {
+            fprintf(stderr, "ERROR: cannot open log file for writing\n");
+            return 1;
+        }
+        fprintf(g_log_file,
+                "{\n"
+                "  \"impl\": \"c\",\n"
+                "  \"example\": \"ecg_anomaly_ae\",\n"
+                "  \"config\": {\"epochs\": %d, \"batch\": %d, \"lr\": %.6f, "
+                "\"momentum\": %.6f, \"seed\": %d, \"shuffle_seed\": %d},\n"
+                "  \"epochs\": [\n",
+                EPOCHS, BATCH, (double)LR, (double)MOMENTUM, SEED, SHUFFLE_SEED);
+        fflush(g_log_file);
+
+        clock_gettime(CLOCK_MONOTONIC, &g_epoch_t0);
+
+        trainingRunResult_t result = trainingRun(
+            model, MODEL_SIZE,
+            (lossConfig_t){
+                .funcType = MSE, .backwardReduction = REDUCTION_MEAN, .classWeights = NULL},
+            trainLoader, valLoader, sgd, EPOCHS, calculateGradsSequential, inferenceWithLoss,
+            epochCallback);
+        (void)result;
+
+        float testLoss =
+            evaluationEpoch(model, MODEL_SIZE, MSE, testLoader, inferenceWithLoss, REDUCTION_MEAN);
+
+        fprintf(g_log_file,
+                "\n  ],\n"
+                "  \"final\": {\"test_loss\": %.6f, \"test_acc\": null, "
+                "\"test_auc\": null}\n"
+                "}\n",
+                (double)testLoss);
+        fclose(g_log_file);
+
+        fprintf(stdout, "FINAL test_loss=%.6f\n", (double)testLoss);
     }
-    fprintf(g_log_file,
-            "{\n"
-            "  \"impl\": \"c\",\n"
-            "  \"example\": \"ecg_anomaly_ae\",\n"
-            "  \"config\": {\"epochs\": %d, \"batch\": %d, \"lr\": %.6f, "
-            "\"momentum\": %.6f, \"seed\": %d, \"shuffle_seed\": %d},\n"
-            "  \"epochs\": [\n",
-            EPOCHS, BATCH, (double)LR, (double)MOMENTUM, SEED, SHUFFLE_SEED);
-    fflush(g_log_file);
-
-    clock_gettime(CLOCK_MONOTONIC, &g_epoch_t0);
-
-    trainingRunResult_t result = trainingRun(
-        model, MODEL_SIZE,
-        (lossConfig_t){.funcType = MSE, .backwardReduction = REDUCTION_MEAN, .classWeights = NULL},
-        trainLoader, valLoader, sgd, EPOCHS, calculateGradsSequential, inferenceWithLoss,
-        epochCallback);
-    (void)result;
-
-    /* Final test-set eval. Use evaluationEpoch (loss-only) to skip the
-     * argmax-based metric pass that would do 140-class accuracy on this AE. */
-    float testLoss =
-        evaluationEpoch(model, MODEL_SIZE, MSE, testLoader, inferenceWithLoss, REDUCTION_MEAN);
-
-    fprintf(g_log_file,
-            "\n  ],\n"
-            "  \"final\": {\"test_loss\": %.6f, \"test_acc\": null, "
-            "\"test_auc\": null}\n"
-            "}\n",
-            (double)testLoss);
-    fclose(g_log_file);
-
-    fprintf(stdout, "FINAL test_loss=%.6f\n", (double)testLoss);
 
     int status = 0;
     int rc = writeAllReconstructions(model, MODEL_SIZE, getTestSample, getTestSize(),
                                      "examples/ecg_anomaly_ae/outputs/c_reconstructions.npy");
     if (rc != 0) {
         fprintf(stderr, "ERROR: c_reconstructions.npy write failed (rc=%d)\n", rc);
-        status = 1;
-    }
-
-    rc = writeAllReconstructions(model, MODEL_SIZE, getTrainSample, getTrainSize(),
-                                 "examples/ecg_anomaly_ae/outputs/c_train_recons.npy");
-    if (rc != 0) {
-        fprintf(stderr, "ERROR: c_train_recons.npy write failed (rc=%d)\n", rc);
         status = 1;
     }
