+++ examples/har_classifier/train_c.c
@@ -8,26 +8,26 @@
 #include <sys/stat.h>
 #include <time.h>
 
-#include "AvgPool1d.h"
 #include "CalculateGradsSequential.h"
 #include "Common.h"
 #include "Conv1dApi.h"
 #include "DataLoader.h"
 #include "DataLoaderApi.h"
-#include "Distributions.h"
 #include "FlattenApi.h"
 #include "InferenceApi.h"
-#include "Kernel.h"
 #include "Layer.h"
+#include "LayerCommon.h"
+#include "LayerQuant.h"
 #include "LinearApi.h"
 #include "LossFunction.h"
-#include "MaxPool1d.h"
 #include "NPYLoaderApi.h"
+#include "Pool1dApi.h"
 #include "Quantization.h"
 #include "QuantizationApi.h"
 #include "ReluApi.h"
 #include "SgdApi.h"
 #include "SoftmaxApi.h"
+#include "StateDictApi.h"
 #include "StorageApi.h"
 #include "Tensor.h"
 #include "TensorApi.h"
@@ -56,24 +56,11 @@
 /* 3 x (Conv1d + ReLU + Pool) + Flatten + Linear + Softmax = 12 layers */
 #define MODEL_SIZE 12
 
-/* ------------------------------------------------------------------------- */
-/* Datasets and dataloader thunks (mirrors example/MnistExperiment.c).       */
-/* ------------------------------------------------------------------------- */
-
 static dataset_t g_trainDataset;
 static dataset_t g_valDataset;
 static dataset_t g_testDataset;
 
-/* Per-sample shape after npyLoad strips the leading N dim is [9, 128] (rank-2)
- * for items and rank-0 (single int32 value) for labels. The C model expects
- * rank-3 inputs [B=1, 9, 128] for Conv1d and rank-1 one-hot float labels [6]
- * for CrossEntropy. We rebuild both at load time. */
-
 static void reshapeItemsAddBatchDim(tensorArray_t *items) {
-    /* items->array[i] currently has shape [9, 128] rank-2. Replace with
-     * [1, 9, 128] rank-3. Data layout is row-major and unchanged, so we only
-     * need to swap the shape header (allocate new dims/order arrays of length 3,
-     * free the old ones). */
     for (size_t i = 0; i < items->size; ++i) {
         tensor_t *t = items->array[i];
         size_t oldRank = t->shape->numberOfDimensions;
@@ -98,10 +85,6 @@ static void reshapeItemsAddBatchDim(tensorArray_t *items) {
 }
 
 static tensorArray_t *buildOneHotLabels(tensorArray_t *intLabels) {
-    /* intLabels->array[i] is a rank-0 int32 tensor (single class index 0..5).
-     * We allocate a brand-new tensorArray_t whose entries are rank-1 float32
-     * one-hot tensors of shape [NUM_CLASSES]. The original int32 array is
-     * left intact (caller still owns it). */
     tensorArray_t *out = reserveMemory(sizeof(tensorArray_t));
     tensor_t **arr = reserveMemory(intLabels->size * sizeof(tensor_t *));
     out->array = arr;
@@ -131,6 +114,7 @@ static tensorArray_t *buildOneHotLabels(tensorArray_t *intLabels) {
 }
 
 static void initDataSets(void) {
+    /* Data path: reuse legacy directory; v2 doesn't duplicate the data. */
     tensorArray_t *trainItems = npyLoad("examples/har_classifier/data/train_x.npy");
     tensorArray_t *trainLabelsRaw = npyLoad("examples/har_classifier/data/train_y.npy");
     reshapeItemsAddBatchDim(trainItems);
@@ -159,7 +143,6 @@ static sample_t *getValSample(size_t id) {
 static sample_t *getTestSample(size_t id) {
     return npyGetSample(&g_testDataset, id);
 }
-
 static size_t getTrainSize(void) {
     return g_trainDataset.items->size;
 }
@@ -170,140 +153,90 @@ static size_t getTestSize(void) {
     return g_testDataset.items->size;
 }
 
-/* ------------------------------------------------------------------------- */
-/* Model parameters (file-static — must outlive buildModel).                 */
-/* ------------------------------------------------------------------------- */
+static void buildModel(layer_t **model, layerQuant_t *lq) {
+    /* Block 1: Conv1d(9->16, K=7, padding=SAME), ReLU, MaxPool(K=2, S=2). */
+    model[0] = conv1dLayerInit(
+        &(conv1dInit_t){
+            .inChannels = IN_CHANNELS, .outChannels = C1_OUT, .kernelSize = C1_K, .padding = SAME},
+        lq);
+    model[1] = reluLayerInit(lq);
+    model[2] = maxPool1dLayerInit(
+        &(maxPool1dInit_t){
+            .kernelSize = 2, .stride = 2, .inputChannels = C1_OUT, .inputLength = LEN_INPUT},
+        lq);
 
-/* Conv1d weights: [Cout, Cin, K]. Bias: [Cout] rank-1 (matches Conv1d.c). */
-static float c1_w_data[C1_OUT * IN_CHANNELS * C1_K];
-static size_t c1_w_dims[3] = {C1_OUT, IN_CHANNELS, C1_K};
-static float c1_b_data[C1_OUT];
-static size_t c1_b_dims[1] = {C1_OUT};
+    /* Block 2 */
+    model[3] = conv1dLayerInit(
+        &(conv1dInit_t){
+            .inChannels = C1_OUT, .outChannels = C2_OUT, .kernelSize = C2_K, .padding = SAME},
+        lq);
+    model[4] = reluLayerInit(lq);
+    model[5] = maxPool1dLayerInit(
+        &(maxPool1dInit_t){
+            .kernelSize = 2, .stride = 2, .inputChannels = C2_OUT, .inputLength = LEN_INPUT / 2},
+        lq);
 
-static float c2_w_data[C2_OUT * C1_OUT * C2_K];
-static size_t c2_w_dims[3] = {C2_OUT, C1_OUT, C2_K};
-static float c2_b_data[C2_OUT];
-static size_t c2_b_dims[1] = {C2_OUT};
+    /* Block 3 */
+    model[6] = conv1dLayerInit(
+        &(conv1dInit_t){
+            .inChannels = C2_OUT, .outChannels = C3_OUT, .kernelSize = C3_K, .padding = SAME},
+        lq);
+    model[7] = reluLayerInit(lq);
+    model[8] = avgPool1dLayerInit(
+        &(avgPool1dInit_t){.kernelSize = LEN_INPUT / 4, .stride = LEN_INPUT / 4}, lq);
 
-static float c3_w_data[C3_OUT * C2_OUT * C3_K];
-static size_t c3_w_dims[3] = {C3_OUT, C2_OUT, C3_K};
-static float c3_b_data[C3_OUT];
-static size_t c3_b_dims[1] = {C3_OUT};
-
-/* Linear weights: [outFeat, inFeat]. Bias: [1, outFeat]. */
-static float fc_w_data[NUM_CLASSES * C3_OUT];
-static size_t fc_w_dims[2] = {NUM_CLASSES, C3_OUT};
-static float fc_b_data[NUM_CLASSES];
-static size_t fc_b_dims[2] = {1, NUM_CLASSES};
-
-static parameter_t *buildParam(distributionType_t dist, float *data, size_t *dims, size_t ndim,
-                               size_t fanIn, size_t fanOut) {
-    quantization_t *q = quantizationInitFloat();
-    tensor_t *p = tensorInitWithDistribution(dist, data, dims, ndim, q, NULL, fanIn, fanOut);
-    tensor_t *g = gradInitFloat(p, NULL);
-    return parameterInit(p, g);
-}
-
-/* MaxPool1d/AvgPool1d have no userApi; we mirror UnitTestMaxPool1d.c, but use
- * reserveMemory for backing storage (since these helpers may run more than once
- * and need addresses that survive across calls). */
-
-static layer_t *buildMaxPool1dLayer(size_t kSize, size_t stride, size_t outC, size_t outLen) {
-    quantization_t *q = quantizationInitFloat();
-
-    kernel_t *kernel = reserveMemory(sizeof(kernel_t));
-    initKernel(kernel, kSize, VALID, /*dilation*/ 1, stride);
-
-    /* Argmax buffer is sized for B=1 (training_batch iterates microbatch-by-
-     * microbatch), shape [1, outC, outLen]. */
-    size_t numArgmax = 1 * outC * outLen;
-    int32_t *argmaxBuf = reserveMemory(numArgmax * sizeof(int32_t));
-    size_t *argmaxDims = reserveMemory(3 * sizeof(size_t));
-    argmaxDims[0] = 1;
-    argmaxDims[1] = outC;
-    argmaxDims[2] = outLen;
-    tensor_t *argmax = tensorInitInt32(argmaxBuf, argmaxDims, 3, NULL);
-
-    maxPool1dConfig_t *cfg = reserveMemory(sizeof(maxPool1dConfig_t));
-    initMaxPool1dConfig(cfg, kernel, argmax, q, q);
-
-    layer_t *layer = reserveMemory(sizeof(layer_t));
-    layerConfig_t *lc = reserveMemory(sizeof(layerConfig_t));
-    layer->type = MAXPOOL1D;
-    lc->maxPool1d = cfg;
-    layer->config = lc;
-    return layer;
-}
-
-static layer_t *buildAvgPool1dLayer(size_t kSize, size_t stride) {
-    quantization_t *q = quantizationInitFloat();
-
-    kernel_t *kernel = reserveMemory(sizeof(kernel_t));
-    initKernel(kernel, kSize, VALID, /*dilation*/ 1, stride);
-
-    avgPool1dConfig_t *cfg = reserveMemory(sizeof(avgPool1dConfig_t));
-    initAvgPool1dConfig(cfg, kernel, q, q);
-
-    layer_t *layer = reserveMemory(sizeof(layer_t));
-    layerConfig_t *lc = reserveMemory(sizeof(layerConfig_t));
-    layer->type = AVGPOOL1D;
-    lc->avgPool1d = cfg;
-    layer->config = lc;
-    return layer;
-}
-
-static void buildModel(layer_t **model) {
-    quantization_t *q1 = quantizationInitFloat();
-    quantization_t *q2 = quantizationInitFloat();
-    quantization_t *q3 = quantizationInitFloat();
-    quantization_t *q4 = quantizationInitFloat();
-
-    /* Block 1: Conv1d(9->16, K=7, padding=SAME), ReLU, MaxPool(K=2,S=2). */
-    kernel_t *k1 = reserveMemory(sizeof(kernel_t));
-    initKernel(k1, C1_K, SAME, 1, 1);
-    parameter_t *c1_w =
-        buildParam(XAVIER_UNIFORM, c1_w_data, c1_w_dims, 3, IN_CHANNELS * C1_K, C1_OUT * C1_K);
-    parameter_t *c1_b = buildParam(ZEROS, c1_b_data, c1_b_dims, 1, 1, C1_OUT);
-    model[0] = conv1dLayerInitLegacy(c1_w, c1_b, k1, q1, q2, q3, q4);
-    model[1] = reluLayerInitLegacy(quantizationInitFloat(), quantizationInitFloat());
-    model[2] = buildMaxPool1dLayer(2, 2, C1_OUT, LEN_INPUT / 2);
-
-    /* Block 2: Conv1d(16->32, K=5, padding=SAME), ReLU, MaxPool(K=2,S=2). */
-    kernel_t *k2 = reserveMemory(sizeof(kernel_t));
-    initKernel(k2, C2_K, SAME, 1, 1);
-    parameter_t *c2_w =
-        buildParam(XAVIER_UNIFORM, c2_w_data, c2_w_dims, 3, C1_OUT * C2_K, C2_OUT * C2_K);
-    parameter_t *c2_b = buildParam(ZEROS, c2_b_data, c2_b_dims, 1, 1, C2_OUT);
-    model[3] =
-        conv1dLayerInitLegacy(c2_w, c2_b, k2, quantizationInitFloat(), quantizationInitFloat(),
-                              quantizationInitFloat(), quantizationInitFloat());
-    model[4] = reluLayerInitLegacy(quantizationInitFloat(), quantizationInitFloat());
-    model[5] = buildMaxPool1dLayer(2, 2, C2_OUT, LEN_INPUT / 4);
-
-    /* Block 3: Conv1d(32->64, K=3, padding=SAME), ReLU, AvgPool(K=32,S=32). */
-    kernel_t *k3 = reserveMemory(sizeof(kernel_t));
-    initKernel(k3, C3_K, SAME, 1, 1);
-    parameter_t *c3_w =
-        buildParam(XAVIER_UNIFORM, c3_w_data, c3_w_dims, 3, C2_OUT * C3_K, C3_OUT * C3_K);
-    parameter_t *c3_b = buildParam(ZEROS, c3_b_data, c3_b_dims, 1, 1, C3_OUT);
-    model[6] =
-        conv1dLayerInitLegacy(c3_w, c3_b, k3, quantizationInitFloat(), quantizationInitFloat(),
-                              quantizationInitFloat(), quantizationInitFloat());
-    model[7] = reluLayerInitLegacy(quantizationInitFloat(), quantizationInitFloat());
-    model[8] = buildAvgPool1dLayer(LEN_INPUT / 4, LEN_INPUT / 4);
-
-    /* Head: Flatten, Linear(64 -> 6), Softmax. */
+    /* Head */
     model[9] = flattenLayerInit();
-    parameter_t *fc_w = buildParam(XAVIER_UNIFORM, fc_w_data, fc_w_dims, 2, C3_OUT, NUM_CLASSES);
-    parameter_t *fc_b = buildParam(ZEROS, fc_b_data, fc_b_dims, 2, 1, NUM_CLASSES);
-    model[10] = linearLayerInitLegacy(fc_w, fc_b, quantizationInitFloat(), quantizationInitFloat(),
-                                      quantizationInitFloat(), quantizationInitFloat());
-    model[11] = softmaxLayerInitLegacy(quantizationInitFloat(), quantizationInitFloat());
+    model[10] =
+        linearLayerInit(&(linearInit_t){.inFeatures = C3_OUT, .outFeatures = NUM_CLASSES}, lq);
+    model[11] = softmaxLayerInit(lq);
 }
 
-/* ------------------------------------------------------------------------- */
-/* Per-epoch JSON log writer + epoch callback.                               */
-/* ------------------------------------------------------------------------- */
+/* Load PyTorch state_dict from per-layer .npy files written by
+ * examples/har_classifier/train_pytorch.py --save-weights.
+ *
+ * Returns 0 on success, non-zero on first missing file. */
+static int loadStateDictFromDir(layer_t **model, const char *weightsDir) {
+    /* Param layer order in model[]: model[0] conv1, model[3] conv2,
+     * model[6] conv3, model[10] fc. 4 entries. */
+    char wPath[256], bPath[256];
+    const char *names[4] = {"conv1", "conv2", "conv3", "fc"};
+    tensor_t *w[4] = {0};
+    tensor_t *b[4] = {0};
+
+    for (int i = 0; i < 4; i++) {
+        snprintf(wPath, sizeof(wPath), "%s/%s.weight.npy", weightsDir, names[i]);
+        snprintf(bPath, sizeof(bPath), "%s/%s.bias.npy", weightsDir, names[i]);
+        /* npyLoadFlat (not npyLoad): a weight file is ONE tensor of shape
+         * [out, in, k] (or [out, in] for fc). npyLoad() slices dim0 (the output
+         * axis) into row tensors, so array[0] is only output channel 0; the
+         * subsequent layerLoadWeights memcpy then runs past that short buffer
+         * into heap garbage — the issue #177 collapse. */
+        w[i] = npyLoadFlat(wPath);
+        b[i] = npyLoadFlat(bPath);
+        if (w[i] == NULL || b[i] == NULL) {
+            fprintf(stderr, "loadStateDictFromDir: missing %s or %s\n", wPath, bPath);
+            return 1;
+        }
+    }
+
+    modelLoadStateDict(
+        model, MODEL_SIZE,
+        (stateDictEntry_t[]){
+            {.name = names[0], .weightData = (float *)w[0]->data, .biasData = (float *)b[0]->data},
+            {.name = names[1], .weightData = (float *)w[1]->data, .biasData = (float *)b[1]->data},
+            {.name = names[2], .weightData = (float *)w[2]->data, .biasData = (float *)b[2]->data},
+            {.name = names[3], .weightData = (float *)w[3]->data, .biasData = (float *)b[3]->data},
+        },
+        4);
+
+    /* modelLoadStateDict copied the data into the layers; release the loaders. */
+    for (int i = 0; i < 4; i++) {
+        freeTensor(w[i]);
+        freeTensor(b[i]);
+    }
+    return 0;
+}
 
 static FILE *g_log_file = NULL;
 static int g_first_epoch = 1;
@@ -353,64 +286,78 @@ int main(void) {
 
     initDataSets();
 
-    dataLoader_t *trainLoader = dataLoaderInit(getTrainSample, getTrainSize, BATCH, NULL, NULL,
-                                               /*shuffle*/ true, /*shuffleSeed*/ SHUFFLE_SEED,
-                                               /*dropLast*/ true);
-    dataLoader_t *valLoader = dataLoaderInit(getValSample, getValSize, 1, NULL, NULL,
-                                             /*shuffle*/ false, /*shuffleSeed*/ 0,
-                                             /*dropLast*/ true);
     dataLoader_t *testLoader = dataLoaderInit(getTestSample, getTestSize, 1, NULL, NULL,
                                               /*shuffle*/ false, /*shuffleSeed*/ 0,
                                               /*dropLast*/ true);
 
+    layerQuant_t lq;
+    layerQuantInitUniform(&lq, quantizationInitFloat());
+
     layer_t *model[MODEL_SIZE];
-    buildModel(model);
+    buildModel(model, &lq);
 
-    optimizer_t *sgd =
-        sgdMCreateOptim(LR, MOMENTUM, /*weightDecay*/ 0.0f, model, MODEL_SIZE, FLOAT32);
+    const char *bitParity = getenv("BIT_PARITY");
+    if (bitParity != NULL && bitParity[0] != '\0') {
+        /* Bit-parity mode: load PyTorch state_dict, skip training, run inference. */
+        const char *wDir = "examples/har_classifier/weights";
+        if (loadStateDictFromDir(model, wDir) != 0) {
+            fprintf(stderr, "BIT_PARITY: state_dict load failed\n");
+            return 1;
+        }
+        fprintf(stdout, "BIT_PARITY: loaded state_dict from %s\n", wDir);
+    } else {
+        dataLoader_t *trainLoader = dataLoaderInit(getTrainSample, getTrainSize, BATCH, NULL, NULL,
+                                                   /*shuffle*/ true, /*shuffleSeed*/ SHUFFLE_SEED,
+                                                   /*dropLast*/ true);
+        dataLoader_t *valLoader = dataLoaderInit(getValSample, getValSize, 1, NULL, NULL,
+                                                 /*shuffle*/ false, /*shuffleSeed*/ 0,
+                                                 /*dropLast*/ true);
 
-    g_log_file = fopen("examples/har_classifier/logs/c.json", "w");
-    if (!g_log_file) {
-        fprintf(stderr, "ERROR: cannot open log file for writing\n");
-        return 1;
+        optimizer_t *sgd =
+            sgdMCreateOptim(LR, MOMENTUM, /*weightDecay*/ 0.0f, model, MODEL_SIZE, FLOAT32);
+
+        g_log_file = fopen("examples/har_classifier/logs/c.json", "w");
+        if (!g_log_file) {
+            fprintf(stderr, "ERROR: cannot open log file for writing\n");
+            return 1;
+        }
+        fprintf(g_log_file,
+                "{\n"
+                "  \"impl\": \"c\",\n"
+                "  \"example\": \"har_classifier\",\n"
+                "  \"config\": {\"epochs\": %d, \"batch\": %d, \"lr\": %.6f, "
+                "\"momentum\": %.6f, \"seed\": %d, \"shuffle_seed\": %d},\n"
+                "  \"epochs\": [\n",
+                EPOCHS, BATCH, (double)LR, (double)MOMENTUM, SEED, SHUFFLE_SEED);
+        fflush(g_log_file);
+
+        clock_gettime(CLOCK_MONOTONIC, &g_epoch_t0);
+
+        trainingRunResult_t result =
+            trainingRun(model, MODEL_SIZE,
+                        (lossConfig_t){.funcType = CROSS_ENTROPY,
+                                       .backwardReduction = REDUCTION_MEAN,
+                                       .classWeights = NULL},
+                        trainLoader, valLoader, sgd, EPOCHS, calculateGradsSequential,
+                        inferenceWithLoss, epochCallback);
+        (void)result;
+
+        epochStats_t testStats = evaluationEpochWithMetrics(
+            model, MODEL_SIZE, CROSS_ENTROPY, testLoader, inferenceWithLoss, REDUCTION_MEAN);
+
+        fprintf(g_log_file,
+                "\n  ],\n"
+                "  \"final\": {\"test_loss\": %.6f, \"test_acc\": %.6f, "
+                "\"test_auc\": null}\n"
+                "}\n",
+                (double)testStats.loss, (double)testStats.accuracy);
+        fclose(g_log_file);
+
+        fprintf(stdout, "FINAL test_loss=%.4f test_acc=%.4f\n", (double)testStats.loss,
+                (double)testStats.accuracy);
     }
-    fprintf(g_log_file,
-            "{\n"
-            "  \"impl\": \"c\",\n"
-            "  \"example\": \"har_classifier\",\n"
-            "  \"config\": {\"epochs\": %d, \"batch\": %d, \"lr\": %.6f, "
-            "\"momentum\": %.6f, \"seed\": %d, \"shuffle_seed\": %d},\n"
-            "  \"epochs\": [\n",
-            EPOCHS, BATCH, (double)LR, (double)MOMENTUM, SEED, SHUFFLE_SEED);
-    fflush(g_log_file);
 
-    clock_gettime(CLOCK_MONOTONIC, &g_epoch_t0);
-
-    trainingRunResult_t result = trainingRun(
-        model, MODEL_SIZE,
-        (lossConfig_t){
-            .funcType = CROSS_ENTROPY, .backwardReduction = REDUCTION_MEAN, .classWeights = NULL},
-        trainLoader, valLoader, sgd, EPOCHS, calculateGradsSequential, inferenceWithLoss,
-        epochCallback);
-    (void)result;
-
-    epochStats_t testStats = evaluationEpochWithMetrics(
-        model, MODEL_SIZE, CROSS_ENTROPY, testLoader, inferenceWithLoss, REDUCTION_MEAN);
-
-    fprintf(g_log_file,
-            "\n  ],\n"
-            "  \"final\": {\"test_loss\": %.6f, \"test_acc\": %.6f, "
-            "\"test_auc\": null}\n"
-            "}\n",
-            (double)testStats.loss, (double)testStats.accuracy);
-    fclose(g_log_file);
-
-    fprintf(stdout, "FINAL test_loss=%.4f test_acc=%.4f\n", (double)testStats.loss,
-            (double)testStats.accuracy);
-
-    /* Predictions: run inference on every test sample, write argmax to .npy.
-     * inference() returns a fresh tensor we own; freeing every iteration via
-     * freeTensor would also free its data buffer, which is what we want. */
+    /* Predictions on test set (both modes). */
     size_t numTest = getTestSize();
     int32_t *predictions = malloc(numTest * sizeof(int32_t));
     if (!predictions) {
